Add support for terabytes in -X options
This fix adds support for 'T' and 't' suffix (indicating terabytes)
to -X options that take <size> parameter.

Made sure that this added support only applies to 64-bit versions of
OpenJ9, including unit tests in openj9/runtime/vm/jvminit.c to check
that up to 2^63-1 size is supported and that 32-bit version of OpenJ9
will not accept this new size suffix.

Fixes: #7427

Signed-off-by: Alexandre Vermeerbergen <avermeerbergen@gmail.com>